Move DefaultPlayerFlags to CampaignData
